Implement `respond_to?` in module Monad
Prompted by the statement “if this was production code, we should remember to implement `#respond_to?` as well” at http://codon.com/refactoring-ruby-with-monads.

I implemented `respond_to?` by implementing `respond_to_missing?`, as recommended by http://robots.thoughtbot.com/always-define-respond-to-missing-when-overriding, so that `method` also works. This requires Ruby 1.9 and later.